refactor: 공백 없애기
