+++ src/App.tsx
@@ -16,7 +16,7 @@ import Add from './pages/add/Add';
 import Splash from './pages/splash/Splash';
 import Chat from './pages/chat/Chat';
 import DocList from './pages/docList/DocList';
-import FormStepper from './pages/\bdocRegister/FormStepper';
+import FormStepper from './pages/docRegister/FormStepper';
 import Result from './pages/result/Result';
 import HosList from './pages/result/KakaoMap/HosList';
 import DocDetail from './pages/docList/DocDetail';
