+++ src/pages/docRegister/FormStepper.tsx
