+++ src/pages/docRegister/component/CheckBox.tsx
