+++ src/pages/docRegister/component/StepNavigation.tsx
