+++ src/pages/docRegister/formSteps/Setp5Treatment.tsx
