+++ src/pages/docRegister/formSteps/Step1Agreement.tsx
