+++ src/pages/docRegister/formSteps/Step6Complete.tsx
