+++ src/pages/docRegister/schemas/Schema.ts
