+++ src/pages/docRegister/types/formTypes.ts
